f1/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
azure-core==1.32.0
azure-functions==1.21.3
azure-storage-blob==12.25.0
blis==0.7.11
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.16.0
confection==0.1.5
cryptography==44.0.2
cymem==2.0.11
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl#sha256=6215d71a3212690e9aec49408a27e3fe6ad7cd6c715476e93d70dc784041e93e
frozenlist==1.5.0
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
MarkupSafe==3.0.2
multidict==6.1.0
murmurhash==1.0.12
numpy==1.26.4
packaging==24.2
preshed==3.0.9
propcache==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
requests==2.32.3
setuptools==76.0.0
six==1.17.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
thinc==8.2.5
tqdm==4.67.1
typer==0.9.4
typing_extensions==4.12.2
urllib3==2.3.0
wasabi==1.1.3
weasel==0.3.4
yarl==1.18.3
